requirements.txt
PyYAML~=6.0.2
prometheus_client~=0.22.0
